fix: render help pages for hidden and 2nd level commands
Added `KartGroup` to 2nd level commands, and added help pages
for hidden commands.
