+++ kart/cli_util.py
@@ -27,6 +27,53 @@ class KartCommand(click.Command):
             return super().format_help(ctx, formatter)
 
 
+class KartGroup(click.Group):
+    command_class = KartCommand
+
+    def get_command(self, ctx, cmd_name):
+        rv = super().get_command(ctx, cmd_name)
+        if rv is not None:
+            return rv
+
+        # typo? Suggest similar commands.
+        import difflib
+
+        matches = difflib.get_close_matches(
+            cmd_name, list(self.list_commands(ctx)), n=3
+        )
+
+        fail_message = f"kart: '{cmd_name}' is not a kart command. See 'kart --help'.\n"
+        if matches:
+            if len(matches) == 1:
+                fail_message += "\nThe most similar command is\n"
+            else:
+                fail_message += "\nThe most similar commands are\n"
+            for m in matches:
+                fail_message += f"\t{m}\n"
+        ctx.fail(fail_message)
+
+    def invoke(self, ctx):
+        if ctx.params.get("post_mortem"):
+            try:
+                return super().invoke(ctx)
+            except Exception:
+                try:
+                    import ipdb as pdb
+                except ImportError:
+                    # ipdb is only installed in dev venvs, not releases
+                    import pdb
+                pdb.post_mortem()
+                raise
+        else:
+            return super().invoke(ctx)
+
+    def format_help(self, ctx, formatter):
+        try:
+            render(ctx.command_path)
+        except Exception as e:
+            return super().format_help(ctx, formatter)
+
+
 def render(command_path: str):
     """Sends output to pager depending on current platform"""
     if platform.system() == "Windows":
